2007-02-15  Roland Levillain  <roland@lrde.epita.fr>
	Have Olena tests pass with ICC 9.1.

	* oln/transforms/dwt.hh (oln::transforms::dwt::dwt): Initalize
	dwt::coeffs in ctor to pacify ICC.
	* tests/utils/tests/stat: Be more tolerant on comparisons of
	floating point values.

	* oln/lrde/efigi/Makefile.am,
	* oln/lrde/ufmt/bin/Makefile.am (AM_CXXFLAGS): Set to
	$(CXXFLAGS_OPTIMIZE) $(CXXFLAGS_STRICT_ERRORS) $(ZLIB_CXXFLAGS)
	-DNDEBUG.
	* oln/lrde/efigi/gaussian.cc (main): Use an eplicit
	double-to-float conversion to pacify ICC.

    prcs2git-id: 10.292

+++ olena/oln/lrde/efigi/gaussian.cc
@@ -7,7 +7,6 @@
 #include <oln/convol/fast_gaussian.hh>
 
 
-
 void usage(char* argv[])
 {
   std::cerr << "usage: " << argv[0]
@@ -25,7 +24,8 @@ int main(int argc, char* argv[])
   if (argc != 4)
     usage(argv);
 
-  float sigma = std::atof(argv[2]);
+  // Use an explicit cast pacify ICC 9.1.
+  float sigma = static_cast<float>(std::atof(argv[2]));
   assert(sigma > 0);
   image2d<float> ima = efigi::load_pfm(argv[1]);
   image2d<float> imaG = convol::fast::gaussian(ima, sigma);
